refactor: index 파일명을 컴포넌트명으로 변경
