+++ src/components/Card/Card.stories.jsx
@@ -1,4 +1,4 @@
-import Card from 'components/Card';
+import Card from 'components/Card/Card';
 
 export default {
   title: 'CardAddition/Card',
