+++ src/containers/CardConfirmModal/CardConfirmModal.jsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
-import PageTitle from 'components/PageTitle';
-import Card from 'components/Card';
-import Form from 'components/Form';
-import NextButton from 'components/NextButton';
-import FlexColumnBox from 'components/FlexColumnBox';
+import PageTitle from 'components/PageTitle/PageTitle';
+import Card from 'components/Card/Card';
+import Form from 'components/Form/Form';
+import NextButton from 'components/NextButton/NextButton';
+import FlexColumnBox from 'components/FlexColumnBox/FlexColumnBox';
 import styled from 'styled-components';
 import ModalOverlay from 'common/Modal/ModalOverlay';
