+++ src/containers/CardCvcInput/CardCvc.jsx
@@ -4,11 +4,11 @@ import validator from 'lib/validations';
 import { CARD_COMPANIES } from 'lib/constants';
 import { CardDispatchContext, CardStateContext } from 'store/card/CardContext';
 import { TYPES } from 'store/card/types';
-import Label from 'components/Label';
-import InputContainer from 'components/InputContainer';
-import InputBox from 'components/InputBox';
-import Tooltip from 'components/Tooltip';
-import { InputBasic } from 'components/Input';
+import Label from 'components/Label/Label';
+import InputContainer from 'components/InputContainer/InputContainer';
+import InputBox from 'components/InputBox/InputBox';
+import Tooltip from 'components/Tooltip/Tooltip';
+import { InputBasic } from 'components/Input/Input';
 
 export default function CardCvc({ onClickTooltip }) {
   const { cardCvc, cardCvcErrorMessage, cardCompanyIndex } = useContext(CardStateContext);
