+++ src/containers/CardExpirationInput/CardExpiration.jsx
@@ -4,10 +4,10 @@ import validator from 'lib/validations';
 import { CardDispatchContext, CardStateContext } from 'store/card/CardContext';
 import { TYPES } from 'store/card/types';
 import AutoFocusInputBox from 'common/AutoFocusInputBox';
-import Label from 'components/Label';
-import InputContainer from 'components/InputContainer';
-import InputBox from 'components/InputBox';
-import { InputBasic } from 'components/Input';
+import Label from 'components/Label/Label';
+import InputContainer from 'components/InputContainer/InputContainer';
+import InputBox from 'components/InputBox/InputBox';
+import { InputBasic } from 'components/Input/Input';
 
 export default function CardExpiration({ color }) {
   const { cardExpiration, cardExpirationErrorMessage } = useContext(CardStateContext);
