+++ src/containers/CardListModal/CardListModal.jsx
@@ -1,5 +1,5 @@
 import { useCallback, useContext } from 'react';
-import CardCompany from 'components/CardCompany';
+import CardCompany from 'components/CardCompany/CardCompany';
 import { CARD_COMPANIES } from 'lib/constants';
 import { CardDispatchContext, CardStateContext } from 'store/card/CardContext';
 import { TYPES } from 'store/card/types';
