+++ src/containers/CardOwnerInput/CardOwner.jsx
@@ -4,11 +4,11 @@ import validator from 'lib/validations';
 import ErrorMessage from 'containers/ErrorMessage/ErrorMessage';
 import { CardDispatchContext, CardStateContext } from 'store/card/CardContext';
 import { TYPES } from 'store/card/types';
-import Label from 'components/Label';
-import InputContainer from 'components/InputContainer';
-import InputBox from 'components/InputBox';
+import Label from 'components/Label/Label';
+import InputContainer from 'components/InputContainer/InputContainer';
+import InputBox from 'components/InputBox/InputBox';
 import styled from 'styled-components';
-import { InputBasicLeft } from 'components/Input';
+import { InputBasicLeft } from 'components/Input/Input';
 
 export default function CardOwner({ color }) {
   const { cardOwner, cardOwnerErrorMessage } = useContext(CardStateContext);
