+++ src/containers/TipModal/TipModal.jsx
