+++ src/pages/CardList.jsx
@@ -3,15 +3,15 @@ import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 import { TYPES } from 'store/card/types';
 import { CardDispatchContext, CardStateContext } from 'store/card/CardContext';
-import PageTitle from 'components/PageTitle';
-import AnotherCard from 'components/AnotherCard';
-import Card from 'components/Card';
-import FlexColumnBox from 'components/FlexColumnBox';
-import Container from 'components/Container';
+import PageTitle from 'components/PageTitle/PageTitle';
+import AnotherCard from 'components/AnotherCard/AnotherCard';
+import Card from 'components/Card/Card';
+import FlexColumnBox from 'components/FlexColumnBox/FlexColumnBox';
+import Container from 'components/Container/Container';
 import PointerBox from 'components/PointerBox/PointerBox';
 import DroppableArea from 'common/DragDrop/DroppableArea';
 import DraggableCard from 'common/DragDrop/DraggableCard';
-import CardConfirmModal from 'containers/CardConfirmModal';
+import CardConfirmModal from 'containers/CardConfirmModal/CardConfirmModal';
 import ClickCardBox from 'common/ClickCardBox';
 import CardManageModal from 'containers/CardManageModal/CardManageModal';
